refactor: make customer portal extend customer api
