--- src/elements/internal/InternalDemoCustomerApi/InternalDemoCustomerApi.ts
@@ -1,9 +1,0 @@
-import { API } from '@foxy.io/sdk/customer';
-import { CustomerApi } from '../../public/CustomerApi';
-import { InternalDemoAPI } from './InternalDemoAPI';
-
-export class InternalDemoCustomerApi extends CustomerApi {
-  get api(): API {
-    return new InternalDemoAPI(this) as API;
-  }
-}
