--- src/elements/internal/InternalDemoCustomerApi/index.ts
@@ -1,5 +1,0 @@
-import { InternalDemoCustomerApi } from './InternalDemoCustomerApi';
-
-customElements.define('foxy-internal-demo-customer-api', InternalDemoCustomerApi);
-
-export { InternalDemoCustomerApi };
