+++ src/elements/public/CustomerApi/CustomerApi.ts
@@ -1,6 +1,7 @@
-import { LitElement, PropertyDeclarations } from 'lit-element';
+import { LitElement, PropertyDeclarations, TemplateResult, html } from 'lit-element';
 
 import { API } from '@foxy.io/sdk/customer';
+import { ConfigurableMixin } from '../../../mixins/configurable';
 import { FetchEvent } from '../NucleonElement/FetchEvent';
 
 /**
@@ -12,118 +13,88 @@ import { FetchEvent } from '../NucleonElement/FetchEvent';
  * @element foxy-customer-api
  * @since 1.4.0
  */
-export class CustomerApi extends LitElement {
+export class CustomerApi extends ConfigurableMixin(LitElement) {
   static readonly SignOutEvent = class extends CustomEvent<void> {};
 
   static readonly SignInEvent = class extends CustomEvent<void> {};
 
   static get properties(): PropertyDeclarations {
     return {
-      storage: { type: String },
-      level: { type: Number },
-      base: { type: String },
+      ...super.properties,
+      storage: { type: String, noAccessor: true },
+      level: { type: Number, noAccessor: true },
+      base: { type: String, noAccessor: true },
     };
   }
 
-  /** Credentials storage implementing Web Storage API. Access tokens and other related info will be stored here. Defaults to in-memory storage. */
-  storage: 'memory' | 'session' | 'local' | Storage = 'memory';
+  private __storage: 'memory' | 'session' | 'local' = 'local';
 
-  /** Numeric Consola log level. If omitted, Consola defaults will be used. */
-  level = 0;
+  private __level = 0;
 
-  /** Bookmark URL for this API. This is where the tree traversal begins. We also use this URL as a base for relative paths. */
-  base = '';
+  private __base = location.origin;
 
-  private __handleFetch = async (evt: unknown) => {
-    if (!(evt instanceof FetchEvent) || evt.defaultPrevented) return;
-    if (evt.eventPhase === evt.AT_TARGET) return;
+  private __api: API = this.__createAPIInstance();
 
-    const getResponse = async () => {
-      console.log(evt.request);
-      if (evt.request.url === 'foxy://auth/session') {
-        try {
-          if (evt.request.method === 'DELETE') {
-            await this.api.signOut();
-            this.dispatchEvent(new CustomerApi.SignOutEvent('signout'));
-            return new Response();
-          }
+  private __handleFetch = async (originalEvent: Event) => {
+    if (!(originalEvent instanceof FetchEvent)) return;
+    if (originalEvent.defaultPrevented) return;
+    if (originalEvent.composedPath()[0] === this) return;
 
-          if (evt.request.method === 'POST') {
-            const payload = await evt.request.clone().json();
+    originalEvent.preventDefault();
 
-            if (payload.type === 'password') {
-              await this.api.signIn(payload.credential);
-              this.dispatchEvent(new CustomerApi.SignInEvent('signin'));
-            }
+    const apiEvent = new FetchEvent('fetch', {
+      cancelable: true,
+      composed: true,
+      request: originalEvent.request,
+      bubbles: true,
+      reject: (reason: unknown) => originalEvent.respondWith(Promise.reject(reason)),
+      resolve: (response: Response) => {
+        if (apiEvent.request.url === 'foxy://customer-api/session') this.requestUpdate();
+        if (response.status === 401) this.requestUpdate();
+        originalEvent.respondWith(Promise.resolve(response));
+      },
+    });
 
-            return new Response(
-              JSON.stringify({
-                _links: { self: { href: evt.request.url } },
-                ...payload,
-              })
-            );
-          }
-        } catch (err) {
-          const status = err instanceof API.AuthError && err.code === 'UNAUTHORIZED' ? 401 : 500;
-          return new Response(null, { status });
-        }
-      }
-
-      if (evt.request.url === 'foxy://auth/recover') {
-        try {
-          if (evt.request.method === 'POST') {
-            const payload = await evt.request.clone().json();
-
-            if (payload.type === 'email') {
-              await this.api.sendPasswordResetEmail(payload.detail);
-              return new Response(
-                JSON.stringify({
-                  _links: { self: { href: evt.request.url } },
-                  email: payload.email,
-                })
-              );
-            }
-          }
-        } catch {
-          //
-        }
-
-        return new Response(null, { status: 500 });
-      }
-
-      const response = await this.api.fetch(evt.request);
-
-      if (response.status === 401) {
-        await this.api.signOut();
-        this.dispatchEvent(new CustomerApi.SignOutEvent('signout'));
-      }
-
-      return response;
-    };
-
-    evt.respondWith(getResponse());
+    if (this.dispatchEvent(apiEvent)) apiEvent.respondWith(this.__handleRequest(apiEvent.request));
   };
+
+  get isLoggedIn(): boolean {
+    return !!this.api.storage.getItem(API.SESSION);
+  }
 
   /** `FoxySDK.Customer.API` instance used by this element to communicate with the backend. */
   get api(): API {
-    const storage =
-      typeof this.storage === 'object'
-        ? this.storage
-        : this.storage === 'local'
-        ? localStorage
-        : this.storage === 'session'
-        ? sessionStorage
-        : undefined;
-
-    return new API({
-      level: this.level,
-      base: new URL(this.base),
-      ...(storage ? { storage } : undefined),
-    });
+    return this.__api;
   }
 
-  createRenderRoot(): CustomerApi {
-    return this;
+  /** Bookmark URL for this API. This is where the tree traversal begins. We also use this URL as a base for relative paths. */
+  get base(): string {
+    return this.__base;
+  }
+
+  set base(value: string) {
+    this.__base = value;
+    this.__api = this.__createAPIInstance();
+  }
+
+  /** Numeric Consola log level. If omitted, Consola defaults will be used. */
+  get level(): number {
+    return this.__level;
+  }
+
+  set level(value: number) {
+    this.__level = value;
+    this.__api = this.__createAPIInstance();
+  }
+
+  /** Credentials storage implementing Web Storage API. Access tokens and other related info will be stored here. Defaults to in-memory storage. */
+  get storage(): 'memory' | 'session' | 'local' {
+    return this.__storage;
+  }
+
+  set storage(value: 'memory' | 'session' | 'local') {
+    this.__storage = value;
+    this.__api = this.__createAPIInstance();
   }
 
   connectedCallback(): void {
@@ -131,8 +102,105 @@ export class CustomerApi extends LitElement {
     this.addEventListener('fetch', this.__handleFetch);
   }
 
+  render(): TemplateResult {
+    return html`
+      ${this._renderTemplateOrSlot(this.isLoggedIn ? 'logged-in' : 'logged-out')}
+      <slot></slot>
+    `;
+  }
+
   disconnectedCallback(): void {
     super.disconnectedCallback();
     this.removeEventListener('fetch', this.__handleFetch);
+  }
+
+  private __createAPIInstance() {
+    const storage =
+      typeof this.__storage === 'object'
+        ? this.__storage
+        : this.__storage === 'local'
+        ? localStorage
+        : this.__storage === 'session'
+        ? sessionStorage
+        : undefined;
+
+    return new API({
+      level: this.__level,
+      base: new URL(this.__base),
+      ...(storage ? { storage } : undefined),
+    });
+  }
+
+  private async __handleVirtualAuthSessionDelete() {
+    await this.api.signOut();
+
+    this.dispatchEvent(new CustomerApi.SignOutEvent('signout'));
+    this.requestUpdate();
+
+    return new Response();
+  }
+
+  private async __handleVirtualAuthSessionPost(request: Request) {
+    const payload = await request.clone().json();
+
+    if (payload.type === 'password') {
+      await this.api.signIn(payload.credential);
+    } else {
+      throw new Error(`Unknown payload type ${payload.type}`);
+    }
+
+    this.dispatchEvent(new CustomerApi.SignInEvent('signin'));
+    this.requestUpdate();
+
+    return new Response(
+      JSON.stringify({
+        _links: { self: { href: request.url } },
+        ...payload,
+      })
+    );
+  }
+
+  private async __handleVirtualAuthRecoverPost(request: Request) {
+    const payload = await request.clone().json();
+
+    if (payload.type === 'email') {
+      await this.api.sendPasswordResetEmail(payload.detail);
+      return new Response(
+        JSON.stringify({
+          _links: { self: { href: request.url } },
+          email: payload.email,
+        })
+      );
+    } else {
+      throw new Error(`Unknown payload type ${payload.type}`);
+    }
+  }
+
+  private async __handleRequest(request: Request) {
+    const url = request.url;
+    const method = request.method;
+
+    try {
+      if (url === 'foxy://customer-api/session') {
+        if (method === 'DELETE') return this.__handleVirtualAuthSessionDelete();
+        if (method === 'POST') return this.__handleVirtualAuthSessionPost(request);
+      }
+
+      if (url === 'foxy://customer-api/recover') {
+        if (method === 'POST') return this.__handleVirtualAuthRecoverPost(request);
+      }
+
+      const response = await this.api.fetch(request);
+
+      if (response.status === 401) {
+        this.api.storage.clear();
+        this.dispatchEvent(new CustomerApi.SignOutEvent('signout'));
+        this.requestUpdate();
+      }
+
+      return response;
+    } catch (err) {
+      return new Response(null, { status: 500 });
+    }
   }
 }
